requirements.txt
asgiref==3.11.1
Django==4.2.30
djangorestframework==3.16.1
numpy==2.0.2
pandas==2.3.3
psycopg==3.2.13
psycopg-binary==3.2.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.2
